Basic adjustments to run LSBQe on new API (incomplete)
